pyproject.toml
[project]
name = "jira-report-generator"
version = "0.1.0"
description = ""
authors = [
    { name = "gab-25", email = "gabrielesorci.25@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jira (>=3.10.5,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]

[project.scripts]
jira-report-generator = 'jira_report_generator.main:run'


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
